Moving data structures into data_structures
